requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
pandas==2.1.4
numpy==1.26.2
fastavro==1.9.3
pyarrow==14.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.25
jupyter==1.0.0
requests==2.31.0
tabulate==0.9.0